Moved RedisClientPool up from cluster package
