+++ modules/examples/src/main/scala/effredis/Cluster.scala
@@ -26,18 +26,18 @@ import log4cats._
 
 object Cluster extends LoggerIOApp {
 
-  val nKeys = 60000
+  val nKeys = 100000
   def program: IO[Unit] =
     RedisClusterClient.make[IO](new URI("http://localhost:7000")).flatMap { cl =>
       for {
         // optionally the cluster topology can be refreshed to reflect the latest partitions
         // this step schedules that job at a pre-configured interval
-        _ <- ClusterUtils.repeatAtFixedRate(10.seconds, cl.topologyCache.expire).start
+        _ <- ClusterUtils.repeatAtFixedRate(4.seconds, cl.topologyCache.expire).start
         _ <- RedisClientPool.poolResource[IO].use { pool =>
               implicit val p = pool
               for {
                 _ <- (0 to nKeys)
-                      .map(i => cl.set(s"key$i", s"value $i"))
+                      .map(i => cl.set(s"dg-key$i", s"value $i"))
                       .toList
                       .sequence
               } yield ()
