+++ modules/tests/src/test/scala/effredis/cluster/TestBenchScenarios.scala
@@ -19,6 +19,7 @@ package effredis.cluster
 import cats.effect._
 import cats.implicits._
 import effredis.Log.NoOp._
+import effredis.RedisClientPool
 
 import effredis.EffRedisFunSuite._
